move REST call into its own resource to hide it for non gitlabci profiles
